requirements.txt
# requirements.txt
flask==2.0.1
werkzeug==2.0.1
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
google-cloud-storage==2.7.0
google-cloud-pubsub==2.13.0
firebase-admin==6.1.0
tensorflow==2.12.0
plotly==5.13.1
joblib==1.2.0
reportlab==3.6.12
seaborn==0.12.2
matplotlib==3.7.1
gunicorn==20.1.0
statsmodels==0.14.1  # Added statsmodels